+++ data_hub_api/docmaps/v1/codecs/elife_manuscript.py
@@ -1,6 +1,6 @@
 from typing import Optional
-from data_hub_api.docmaps.api_input_typing import ApiInput, ApiPreprintInput
-from data_hub_api.docmaps.docmap_typing import (
+from data_hub_api.docmaps.v1.api_input_typing import ApiInput, ApiPreprintInput
+from data_hub_api.docmaps.v1.docmap_typing import (
     DocmapAssertionItem,
     DocmapElifeManuscriptOutput
 )
