+++ data_hub_api/docmaps/v1/codecs/evaluation.py
@@ -2,14 +2,14 @@ from datetime import datetime
 import logging
 from typing import Iterable, Optional, Sequence, cast, Tuple
 
-from data_hub_api.docmaps.codecs.elife_manuscript import get_elife_manuscript_version_doi
-from data_hub_api.docmaps.api_input_typing import (
+from data_hub_api.docmaps.v1.codecs.elife_manuscript import get_elife_manuscript_version_doi
+from data_hub_api.docmaps.v1.api_input_typing import (
     ApiEditorDetailInput,
     ApiEvaluationInput,
     ApiInput,
     ApiPreprintInput
 )
-from data_hub_api.docmaps.docmap_typing import (
+from data_hub_api.docmaps.v1.docmap_typing import (
     DocmapAction,
     DocmapContent,
     DocmapEvaluationInput,
