+++ data_hub_api/docmaps/v1/sql/docmaps_index.sql
